+++ src/Anchor/index.ts
@@ -1,0 +1,4 @@
+import Anchor from './anchor';
+
+
+export default Anchor;