+++ src/FloatingActionButton/index.ts
@@ -1,0 +1,4 @@
+import FloatingActionButton from './floating_action_button';
+
+
+export default FloatingActionButton;